requirements.txt
openai
tiktoken
rouge
torch
transformers
accelerate
evaluate
xopen
python-dotenv
tqdm
requests
retrying
FlagEmbedding
modelscope
llama-index
llama-index-core
llama-index-embeddings-huggingface
llama-index-embeddings-openai
llama-index-postprocessor-flag-embedding-reranker
llama-index-retrievers-bm25
llama-index-llms-dashscope
llama-index-indices-managed-dashscope